+++ brave/src/test/java/com/linecorp/armeria/common/brave/RequestContextCurrentTraceContextTest.java
@@ -32,7 +32,7 @@ import org.mockito.junit.MockitoRule;
 import org.mockito.stubbing.Answer;
 
 import com.linecorp.armeria.common.RequestContext;
-import com.linecorp.armeria.common.brave.RequestContextCurrentTraceContext.PingPongExtra;
+import com.linecorp.armeria.internal.brave.TraceContextUtil.PingPongExtra;
 import com.linecorp.armeria.common.util.SafeCloseable;
 import com.linecorp.armeria.internal.DefaultAttributeMap;
 import com.linecorp.armeria.internal.brave.TraceContextUtil;
@@ -187,7 +187,7 @@ public class RequestContextCurrentTraceContextTest {
         final TraceContext context = TraceContext.newBuilder().traceId(1).spanId(1)
                                                  .extra(Collections.singletonList(extra)).build();
 
-        PingPongExtra.maybeSetPong(context);
+        TraceContextUtil.PingPongExtra.maybeSetPong(context);
 
         assertThat(extra.isPong()).isTrue();
     }
